+++ SwitchStartupProject/Configuration/MultiProjectConfigurationProject.cs
@@ -13,7 +13,9 @@ namespace LucidConcepts.SwitchStartupProject
             string workingDirectory,
             bool startProject,
             string startExternalProgram,
-            string startBrowserWithUrl)
+            string startBrowserWithUrl,
+            bool? enableRemoteDebugging,
+            string remoteDebuggingMachine)
         {
             this.NameOrPath = nameOrPath;
             this.CommandLineArguments = commandLineArguments;
@@ -21,6 +23,8 @@ namespace LucidConcepts.SwitchStartupProject
             this.StartProject = startProject;
             this.StartExternalProgram = startExternalProgram;
             this.StartBrowserWithUrl = startBrowserWithUrl;
+            this.EnableRemoteDebugging = enableRemoteDebugging;
+            this.RemoteDebuggingMachine = remoteDebuggingMachine;
         }
 
         public string NameOrPath { get; private set; }
@@ -29,5 +33,7 @@ namespace LucidConcepts.SwitchStartupProject
         public bool StartProject { get; private set; }
         public string StartExternalProgram { get; set; }
         public string StartBrowserWithUrl { get; set; }
+        public bool? EnableRemoteDebugging { get; private set; }
+        public string RemoteDebuggingMachine { get; set; }
     }
 }